requirements.txt
streamlit
selenium
mysql-connector-python
beautifulsoup4
lxml
python-dotenv
